overflows manageds

+++ libft/ft_lstnew.c
@@ -6,13 +6,13 @@
 /*   By: ibenhaim <ibenhaim@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/23 13:46:10 by ibenhaim          #+#    #+#             */
-/*   Updated: 2023/01/20 12:44:01 by ibenhaim         ###   ########.fr       */
+/*   Updated: 2023/02/08 11:55:58 by ibenhaim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-t_list	*ft_lstnew(int content)
+t_list	*ft_lstnew(long long int content)
 {
 	t_list	*new;
